requirements.txt
python-dotenv
vk_api
openai
requests
Pillow
reportlab